requirements.txt
python-dateutil==2.8.2
pytz==2023.3
pyrfc==3.3.1
zeep==4.2.1
requests==2.31.0
urllib3==2.0.4
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
alembic==1.12.1
redis==5.0.1
hiredis==2.2.3
celery==5.3.4
kombu==5.3.4
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
tensorflow==2.15.0
torch==2.1.1
transformers==4.35.2
spacy==3.7.2
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
openai==1.3.7
anthropic==0.7.8
langchain==0.0.340
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
loguru==0.7.2
python-json-logger==2.0.7
prometheus-client==0.19.0
sentry-sdk==1.38.0
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
tqdm==4.66.1
rich==13.7.0
apscheduler==3.10.4
cryptography==41.0.7
pyjwt==2.8.0
passlib==1.7.4
docker==6.1.3
kubernetes==28.1.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
tenacity==8.2.3
aiohttp==3.9.1
lxml==4.9.3
jsonschema==4.20.0